+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+3.2.1
+=====
+
+* (improvement) Change configuration of hosting tier to allow better using it.
+* (improvement) Automatically catch all exceptions in the validate app command.
+
+
 3.2.0
 =====